requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Werkzeug
google-generativeai
python-dotenv